Refactor DelegateVotes.jsx by moving subcomponents to separate files (#17)
* feat: Extract VoteDetailsContainer to its own file

* feat: Extract VoteReason subcomponent to its own f

* feat: Updated src/components/Delegates/DelegateVot

* refactor: moved StandardVoteContainer to its own f

* feat: Extract ApprovalVoteContainer to its own fil

* feat: Updated src/components/Delegates/DelegateVot

* feat: Updated src/components/Delegates/DelegateVot

* fix imports

* rearange

---------
